+++ src/pages/main-page/hooks/use-current-offers.ts
@@ -1,8 +1,13 @@
-import { useAppSelector } from '../../../hooks';
+import { useEffect } from 'react';
+import { useAppDispatch, useAppSelector } from '../../../hooks';
 import { getActiveCity, getOffers } from '../../../store/offers-data/selector';
 import { ServerOffer } from '../../../types/offer';
+import { fetchOffersAction } from '../../../store/api-actions';
+import { offersActions } from '../../../store/offers-data/offers-data';
 
 export function useCurrentOffers() {
+	const dispatch = useAppDispatch();
+	dispatch(offersActions.setActiveOffer(null));
 	const offers = useAppSelector(getOffers);
 	const activeCity = useAppSelector(getActiveCity);
 	const offersByCities: Record<string, ServerOffer[]> = {};
@@ -18,6 +23,10 @@ export function useCurrentOffers() {
 	}
 
 	const currentOffers = offersByCities[activeCity];
+
+	useEffect(() => {
+		dispatch(fetchOffersAction());
+	}, [dispatch]);
 
 	return { currentOffers, activeCity };
 }
