+++ src/store/offer-data/offer-data.ts
@@ -16,6 +16,7 @@ const initialState: OfferData = {
 	isReviewsLoading: false,
 	isNearbyLoading: false,
 	isReviewSending: false,
+	hasErrorOfferLoading: false,
 	hasErrorSubmit: false,
 };
 
@@ -27,10 +28,15 @@ export const offerData = createSlice({
 		builder
 			.addCase(fetchFullOfferAction.pending, (state) => {
 				state.isFullOfferLoading = true;
+				state.hasErrorOfferLoading = false;
 			})
 			.addCase(fetchFullOfferAction.fulfilled, (state, action) => {
 				state.fullOffer = action.payload;
 				state.isFullOfferLoading = false;
+			})
+			.addCase(fetchFullOfferAction.rejected, (state) => {
+				state.isFullOfferLoading = false;
+				state.hasErrorOfferLoading = true;
 			})
 			.addCase(fetchReviewsAction.pending, (state) => {
 				state.isReviewsLoading = true;
