+++ src/store/offer-data/selector.ts
@@ -22,5 +22,7 @@ export const getNearbyLoadingStatus = (
 export const getReviewSendingStatus = (
 	state: Pick<State, NameSpace.Offer>
 ): boolean => state[NameSpace.Offer].isReviewSending;
+export const getErrorOfferLoadingStatus = (state: State): boolean =>
+	state[NameSpace.Offer].hasErrorOfferLoading;
 export const getErrorSubmitStatus = (state: State): boolean =>
 	state[NameSpace.Offer].hasErrorSubmit;
